+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 ## [3.1.2]
 - Added new methods to Merchant API `createCheckoutSession` and `updateCheckoutSession`
 - Enhanced interface of the endpoint cardWallet/session to pass create payment request params
+- Added `sessionId` for for `MerchantAPI` operations:
+    - `createPaymentRequest`
+    - `reservation`
 
 ## [3.0.3]